+++ example/lib/main.dart
@@ -10,6 +10,7 @@ import 'package:mapbox_maps_example/ornaments_example.dart';
 import 'package:mapbox_maps_example/geojson_line_example.dart';
 import 'package:mapbox_maps_example/image_source_example.dart';
 import 'package:mapbox_maps_example/map_interface_example.dart';
+import 'package:mapbox_maps_example/interactive_features_example.dart';
 import 'package:mapbox_maps_example/polygon_annotations_example.dart';
 import 'package:mapbox_maps_example/polyline_annotations_example.dart';
 import 'package:mapbox_maps_example/simple_map_example.dart';
@@ -41,6 +42,7 @@ final List<Example> _allPages = <Example>[
   SpinningGlobeExample(),
   FullMapExample(),
   StyleExample(),
+  InteractiveFeaturesExample(),
   CameraExample(),
   ProjectionExample(),
   MapInterfaceExample(),